Refactor to use IBioPolymer and add XML docs
Refactored ClassicSearchEngine and CalibrationTask to use the more generic IBioPolymer type instead of Protein. Added an overloaded constructor to ClassicSearchEngine for List<Protein> to List<IBioPolymer> conversion. Updated IntegerTextBoxControl with XML documentation for OnPreviewTextInput and OnKeyDown methods. Added a call to NewCollection in RnaSearchTask. Updated using directives to include the Omics namespace.